モデルクラスの分割
