+++ project/__init__.py
@@ -18,20 +18,10 @@ login_manager.login_view = 'login'
 # login_manager.init_app(app)
 
 
-# モデルの定義（データベースのテーブル定義）
+# モデルクラスの読み込み
 # **********************************************************************
 
-class User(UserMixin, db.Model):
-    '''
-    Userモデルクラス（usersテーブル）
-    '''
-
-    __tablename__ = 'users'
-    
-    id = db.Column(db.Integer, primary_key=True)
-    email = db.Column(db.String(50), unique=True)
-    password = db.Column(db.String(100))
-    name = db.Column(db.String(20))
+from .models import User
 
 
 # 関数の定義
